Adicionado estrutura do Sphinx com docs em EN do Flask e script de build

+++ gerate_html.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
+
+
 ## Gerate HTML DOCs 
 
 ## Clean folder _build
-rm -rf _build/*
+if [ -d "_build" ]
+then
+	rm -rf _build/*
+	make html
+else
+	mkdir _build
+	make html
+fi
 
-make html
-rsync -av --delete-excluded  _build/html/ ../flask/
+
+if [ -d "../flash" ]
+then
+	rsync -av --delete-excluded  _build/html/ ../flask/
+else
+	mkdir ../flask
+        rsync -av --delete-excluded  _build/html/ ../flask/
+fi
